+++ packages/terra-table/tests/nightwatch/table/TableTestRoutes.jsx
@@ -8,8 +8,6 @@ import NoStripedTable from './components/NoStripedTable';
 import NoPaddingTable from './components/NoPaddingTable';
 import SingleRowTable from './components/SingleRowTable';
 import TableNoRows from './components/TableNoRows';
-import SelectableTable from './components/SelectableTable';
-import SingleRowSelectableTableOnChange from './components/SingleRowSelectableTableOnChange';
 import TableWithHighlightedRows from './components/TableWithHighlightedRows';
 import TableWithSortIndicator from './components/TableWithSortIndicator';
 import TableWithSubheaders from './components/TableWithSubheaders';
@@ -20,13 +18,11 @@ const routes = (
     <Route path="/tests/table-tests/default" component={DefaultTable} />
     <Route path="/tests/table-tests/no-striped" component={NoStripedTable} />
     <Route path="/tests/table-tests/no-padding" component={NoPaddingTable} />
-    <Route path="/tests/table-tests/table-no-rows" component={TableNoRows} />
-    <Route path="/tests/table-tests/single-row-table" component={SingleRowTable} />
-    <Route path="/tests/table-tests/selectable-table-onchange" component={SingleRowSelectableTableOnChange} />
-    <Route path="/tests/table-tests/selectable-table" component={SelectableTable} />
-    <Route path="/tests/table-tests/table-highlighted-rows" component={TableWithHighlightedRows} />
-    <Route path="/tests/table-tests/table-sort-indicator" component={TableWithSortIndicator} />
-    <Route path="/tests/table-tests/table-subheaders" component={TableWithSubheaders} />
+    <Route path="/tests/table-tests/no-rows" component={TableNoRows} />
+    <Route path="/tests/table-tests/single-row" component={SingleRowTable} />
+    <Route path="/tests/table-tests/highlighted-rows" component={TableWithHighlightedRows} />
+    <Route path="/tests/table-tests/sort-indicator" component={TableWithSortIndicator} />
+    <Route path="/tests/table-tests/subheaders" component={TableWithSubheaders} />
   </div>
 );