+++ packages/terra-table/tests/nightwatch/table/components/NoStripedTable.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
-const SingleRowSelectableTable = () => (
-  <Table isStriped={false}>
+const NoStripedTable = () => (
+  <Table isStriped={false} id="Table">
     <Table.Header>
       <Table.HeaderCell content={'Name'} key={'NAME'} minWidth={'small'} />
       <Table.HeaderCell content={'Address'} key={'ADDRESS'} minWidth={'medium'} />
       <Table.HeaderCell content={'Phone Number'} key={'PHONE_NUMBER'} minWidth={'large'} />
     </Table.Header>
-    <Table.SingleSelectableRows>
-      <Table.Row isSelected key={'PERSON_0'}>
+    <Table.Rows>
+      <Table.Row key={'PERSON_0'}>
         <Table.Cell content="John Smith" key="NAME" />
         <Table.Cell content="123 Adams Drive" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
@@ -24,8 +24,8 @@ const SingleRowSelectableTable = () => (
         <Table.Cell content="213 Raymond Road" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
       </Table.Row>
-    </Table.SingleSelectableRows>
+    </Table.Rows>
   </Table>
 );
 
-export default SingleRowSelectableTable;
+export default NoStripedTable;