+++ packages/terra-table/CHANGELOG.md
@@ -7,7 +7,11 @@ Unreleased
 * SelectableTableRows component
 
 ### Changed
+* Refactor nightwatch tests
 * Updated SingleSelectableRows to utilize the SelectableTableRows component
+
+### Fixed
+* Initial Single Select Table render to accept an isSelected Row
 
 1.12.0 - (September 26, 2017)
 ------------------